Read the shipped translations from a support class
The language sweep carried its own file handling: a glob over l10n/, a json_decode,
a filter for plural arrays and a path built with dirname(__DIR__, 3). None of that
is what the test is about, and the path depended on how deep the test file sat.

tests/Support/ShippedTranslations.php owns it now and serves the languages as data
sets directly, so the test declares DataProviderExternal and says
`ShippedTranslations::of($language)` where it needs the strings. What is left in
the test class is the claim it makes and the objects it needs to make it.
